### 2020-11-25 #2865 @whedon
[PRE REVIEW]: pygamma-agreement: Gamma measure for inter/intra-annotator agreement in Python
**Submitting author:** @hadware (<a href="http://orcid.org/0000-0002-8511-1644">Hadrien Titeux</a>)
**Repository:** <a href="https://github.com/bootphon/pygamma-agreement/" target ="_blank">https://github.com/bootphon/pygamma-agreement/</a>
**Version:** 0.1.4
