### 2020-11-25 ff6b09c
Added unit tests for elan and tg loading. Fixed some issues with TextGrid. Added some options to both loading methods.
