### 2020-11-25 8735125
Dropped mandatory dependency for matplotlib.
