### 2021-01-25 #2865 @faroit
@arfon I'm happy to review this contribution. I will have the bandwidth to do it rather soon. Thanks for considering me, the package sounds really interesting and useful!
