### 2021-01-26 #2989 @whedon
[REVIEW]: pygamma-agreement: Gamma measure for inter/intra-annotator agreement in Python
**Submitting author:** @hadware (<a href="http://orcid.org/0000-0002-8511-1644">Hadrien Titeux</a>)
**Repository:** <a href="https://github.com/bootphon/pygamma-agreement/" target ="_blank">https://github.com/bootphon/pygamma-agreement/</a>
**Version:** v0.1.6
