### 2021-01-28 #2865 @arfon
:wave: @apiad @pzelasko @ajoer - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is _pygamma-agreement: Gamma measure for inter/intra-annotator agreement in Python_, with the Python-based software here: https://github.com/bootphon/pygamma-agreement/

Many thanks!
