### 2021-02-03 #2865 @arfon
> I'd be more than happy to help review it! What is the deadline?

We aim to have reviews completed in ~6 weeks. As JOSS reviews are generally iterative in nature, we encourage reviewers to provide initial feedback sooner than this though in case any clarifications are necessary.

Does this sound OK?
