### 2021-02-09 bootphon/pygamma-agreement#18 @faroit
[JOSS] add contribution guidelines
The [JOSS review](https://github.com/openjournals/joss-reviews/issues/2989) criteria asks for "clear guidelines for third-parties wishing to"

* Contribute to the software
* Report issues or problems with the software
* Seek support

ideally, add this to a `contributing.md`
