### 2021-02-09 eebba42
Added a CONTRIBUTING.md file, as per @faroit 's suggestion in #18
