+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+
+# How to Contribute?
+
+## In case you have an issue with `pygamma-agreement`
+
+As expected, submit an issue on this repository and:
+
+* Make sure you provide a clear description of what you were trying to achieve
+* If possible, also provide a working minimal example of code that can reproduce your error
+* Give details on what kind of platform you were running your code on, namely **OS** and **Python version**
+
+## Needing help on using `pygamma-agreement`
+
+If you're trying to do something that isn't clearly explained in our documentation, do not hesitate on 
+seeking help through an issue. If possible,
+
+* State what you're willing to achieve
+* Provide a code sample of what you wre trying to do
+* Try pointing out which part of the documentation felt inadequate regarding your original goal
+
+## Code Contributiions to `pygamma-agreement`
+
+If you're willing to take it upon yourself to improve `pygamma-agreement`, via bugfixes, improvements and new features 
+(which we strongly encourage you to do and are very grateful), please follow these steps:
+
+1. Submit an issue explaining what you're willing to fix or add to this package. We can discuss with you on the 
+   the best way to do it, considering the current state of things.
+2. Fork the code, open a pull-request (PR) and explain what you'll be adding/changing in the description.
+3. Code away, we'll be tracking your progress and commenting on your improvements to guide as best we can.
+   In case you're implementing something new, or fixing a bug that slipped through our testing suite, 
+   please don't think twice on writing new tests.
+4. Once you think you're finished and all the tests are passing, we'll merge, and voilà!
+
