### 2021-02-09 bootphon/pygamma-agreement#18 @hadware
Thanks for that suggestion!

I added a CONTRIBUTING file to the repo, i hope it fits the bill. Do tell me if something isn't right.
