### 2021-02-10 4640384
Update .readthedocs.yml
Bumping python version for the doc because of cvxpy causing some troubles
