+++ .readthedocs.yml
@@ -13,7 +13,7 @@ formats: all
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: 3.6
+  version: 3.7
   install:
     # this runs `pip install .` before the doc is built
     - method: pip
