### 2021-02-11 5194ec2
Fixes to the alignment checking code (used to bug when the continuum contained the same segment annotated twice with two different classes)
