### 2021-02-11 2b550c2
Fixing unit tests for alignment checking
