### 2021-02-11 9cb94f0
Documentation fix for csv loading, bumping version number
