### 2021-02-17 bootphon/pygamma-agreement#20 @apiad
Warning from deprecated use of `numpy` builtin types.
Greetings. I'm one of the reviewers assigned to the submission to JOSS (https://github.com/openjournals/joss-reviews/issues/2989).

I'm in the process of downloading the package and running the tests. All tests worked correctly, however, there are several deprecation warnings from `numpy`.

I'm opening this issue basically to understand what could be the impact of these warnings, if they are easy to fix, or if they are irrelevant.
