### 2021-02-17 bootphon/pygamma-agreement#20 @hadware
hey @apiad , thanks again for reviewing this package.

I'll update my local numpy version to see if I get the same warnings. In the meantime, can you specify your numpy version as to make sure that we are on the same page?
