### 2021-02-17 #2989 @apiad
@arfon I wanted to ask about the depth expected in the technical description of the software. This is the first paper I've reviewed for JOSS and from what I see, papers are usually not expected to go too deeply into the technical or mathematical details of whatever algorithms are implemented, at least not to the same extent as research papers. Am I correct?
