### 2021-02-17 bootphon/pygamma-agreement#20 @apiad
Sure thing! I just pip installed on a fresh environment. It installed `numpy==1.20.1`.
