### 2021-02-17 bootphon/pygamma-agreement#20 @apiad
This is one of the messages:

```tests/test_gamma.py: 18060 warnings
  /home/apiad/.local/lib/python3.8/site-packages/cvxpy/interface/numpy_interface/ndarray_interface.py:47: DeprecationWarning: `np.complex` is a deprecated alias for the builtin `complex`. To silence this warning, use `complex` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.complex128` here.
  Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
    if result.dtype in [numpy.complex, numpy.float64]:
```

Again, I don't really think this is extremely important, not at all, but since we are supposed to verify the functionality claims, just wanted to double check if you are OK with these warnings.
