### 2021-02-17 bootphon/pygamma-agreement#20 @hadware
Alright, I did my updates, and they're all coming from numpy usages from `numba`, `cvxpy` and `ecos` (dependency of cvxpy). It's ugly, but there's not much i can do :man_shrugging: .

Do you think i should silence warnings (like this  https://docs.python.org/3/library/warnings.html#temporarily-suppressing-warnings ) ?
