### 2021-02-17 bootphon/pygamma-agreement#20 @apiad
I'm not sure. If those warnings will appear to the users of your API and you are positive you understand why they are happening, I think it might be a good idea to silence it, so as to not disturb the user with "issues" that they do not understand and can't do anything about. But silencing warnings is always complicated, IMHO, because something important might slip under the radar.

Anyway, on my side, I think I'm pretty much convinced these warnings are not relevant to the review process as they do not hinder in any way the functionalities of the library, so I'll be closing this issue.

Thanks!
