### 2021-02-18 #2989 @Rachine
Hello, I just noticed these last exchanges about the algorithms we used in our paper.
We have slides trying to explain with simple vizualisations what the core algorithm is doing for the gamma agreement.  If it is ok for the review process at JOSS @arfon, happy to share them with you @apiad.
