### 2021-02-22 #2989 @apiad
> @apiad - yes that's correct. We need reviewers to verify the functionality of the software and that can happen in a number of ways including some combination of:
>
> * Inspecting and running automated tests.
> * Executing test procedures based on test inputs/expected outputs.
> * Reviewing the code.
>
> A _deep_ algorithmic review is not generally required but we do want you to be familiar with what the code is doing. Does that sound OK?

Thanks! This is very helpful.
