### 2021-02-23 #2989 @Rachine
> > Hello, I just noticed these last exchanges about the algorithms we used in our paper.
> > We have slides trying to explain with simple vizualisations what the core algorithm is doing for the gamma agreement. If it is ok for the review process at JOSS @arfon, happy to share them with you @apiad.
>
> Yes, that would be very helpful!

Here is the link to the slides: https://docs.google.com/presentation/d/1dIyx_Xs8oN-9xKE4pQin1EdMN7dD90ga1LaSluZ_mps/edit?usp=sharing
If we need to clarify something, or if you need anything do not hesitate to ping us.

> I have already confirmed the functionality claims (by running all tests offline), I'm in the process of running a few manual examples to see the library in action and also check out the performance a little bit. A brief overview of the algorithm's background would be very nice!

Best,
