### 2021-02-23 #2989 @apiad
Thanks, @Rachine, this was very enlightening.

Outside the scope of the submission, I have a question of my own: is there a rule-of-thumb dataset-independent scale for evaluating how good the agreement is?

E.g., in Kappa, I believe if you are somewhere above 0.7 it is considered a "good enough" agreement. Of course these values are always going to be very subjective, but in the end, measuring what is "good enough" is always gonna rest on some degree of subjectivity.

I guess my questions go along the line of is it possible to give the user a qualitative measure of the quality of the agreement, even if it's highly subjective or is this totally dependent on each dataset, and in that case, are there rules-of-thumb that can be used to determine these levels of quality for a specific dataset?
