### 2021-02-23 #2989 @Rachine
> Thanks, @Rachine, this was very enlightening.
>
> Outside the scope of the submission, I have a question of my own: is there a rule-of-thumb dataset-independent scale for evaluating how good the agreement is?

This is a very interesting and important question! We have been asked this question by linguists and speech pathologists also.
The full answer is that we know that only by experience, as the kappa, and this has not been properly measured with psychophysics.  We are launching a collaboration with psychophysics experts, with more experience in perceptual differences to obtain this scale for the Gamma.

For example, we found that to get 0.6-0.8 in practice is really good for real annotations of speech corpora of turn-takings.

Above 0.8 was never found so far for speech annotations, such as type of vocalization; speaker turns.

For instance, we found out that Child-directed speech in Non-native language was harder to annotate.

Maybe, we should gather a table with all Gamma measurements that have been done so far, so that can people can contribute or be able to compare.

>
> E.g., in Kappa, I believe if you are somewhere above 0.7 it is considered a "good enough" agreement. Of course these values are always going to be very subjective, but in the end, measuring what is "good enough" is always gonna rest on some degree of subjectivity.
>
> I guess my questions go along the line of is it possible to give the user a qualitative measure of the quality of the agreement, even if it's highly subjective or is this totally dependent on each dataset, and in that case, are there rules-of-thumb that can be used to determine these levels of quality for a specific dataset?
