### 2021-02-24 #2989 @apiad
@Rachine thanks again for the explanation! I think it would be a very good idea to make these insights public to your library's users, dunno if with some comments in the docs or a repository where you collect these measurements, as you suggested.
