### 2021-02-24 #2989 @apiad
@arfon taking into consideration your comments, my exchanges with the author, and having checked the functional claims and read the submission paper and the library source code (which is conveniently commented and suitably organized), I consider this submission meets the quality standards for publishing in JOSS.

I sincerely appreciate the opportunity to collaborate, and please don't hesitate to contact me again in the future if you consider it helpful.

Big congrats to the authors for the great work done with this library.
