### 2021-02-24 bootphon/pygamma-agreement#21 @Rachine
Independent scale for evaluating how good a given agreement is
As mentionned in the joss reviews here https://github.com/openjournals/joss-reviews/issues/2989#issuecomment-784234602, it would be super valuable to compare, and evaluate with a perceptual scale the values of the gamma.

Basicall `How good is an observed gamma for a given task?`

We can start in the documentation to gather all the gamma that have been measured along the tasks, annotators informations, language, modality.

In the future, we rwill eport here the psychophysical experiment to validate more carefully the gamma.
