### 2021-04-27 #2989 @hadware
Hey @faroit, I see that some of your review points for our library still remain unsolved. Do you need more information from us to be able to resolve them? Do you see any improvement to our work that you think are essential for it to be on par with JOSS's standards?
