### 2021-05-26 fcfcfe7
fixed plot (wrong label order) of categorical dissim. matrix
