### 2021-05-26 a7e3a76
replacement of dissim. matrix argument with a function (tbd if definitive)
