### 2021-05-26 d7f8e45
removal of random set iteration of continuum.categories()
