### 2021-05-27 9c32e2a
adapting tests for non-set format of categories list
