### 2021-05-28 f7d9943
parellellization of random sample disorder computations (40s -> 13s for 3by100)
