### 2021-05-28 0f333c6
documentation for categorical dissimilarity with functions
