+++ pygamma_agreement/dissimilarity.py
@@ -80,7 +80,10 @@ class AbstractDissimilarity:
             return self.build_arrays_alignment(resource),
 
     @staticmethod
-    def tuples_disorders(*args, **kwargs):
+    def alignments_disorders(*args, **kwargs):
+        """
+        Fast computation of alignments disorder using numba.
+        """
         raise NotImplemented()
 
     def __call__(self, *args) -> np.ndarray:
@@ -97,22 +100,29 @@ class CategoricalDissimilarity(AbstractDissimilarity):
     cat_dissimilarity_matrix : optional, f: (str,str) -> float function representing
         the matrix containing the values between categories. Has to be symetrical (f(x, y) = f(y, x))
         with an empty diagonal (f(x, x) = 0). Defaults to setting all dissimilarities to 1.
+        OR
+        N,N matrix (np.array) containing dissimilarity values between categories (works best with ordinal
+        categories, since categories are indexed in alphabetical order).
     delta_empty : optional, float
         empty dissimilarity value. Defaults to 1.
     """
 
     def __init__(self,
                  categories: Iterable[str],
-                 cat_dissimilarity_matrix: Callable[[str, str], float] = (lambda cat1, cat2: float(cat1 != cat2)),
+                 cat_dissimilarity_matrix: Union[Callable[[str, str], float], np.ndarray] = (
+                         lambda cat1, cat2: float(cat1 != cat2)),
                  delta_empty: float = 1):
         super().__init__(delta_empty)
 
         self.categories_dict = {cat: i for i, cat in enumerate(categories)}
         self.categories_nb = len(self.categories_dict)
-        # TODO: make sure that the categorical dissim matrix matches the categories order
-        self.cat_matrix = np.array([
-            [cat_dissimilarity_matrix(cat1, cat2) for cat1 in categories]
-            for cat2 in categories])
+
+        if isinstance(cat_dissimilarity_matrix, np.ndarray):
+            self.cat_matrix = cat_dissimilarity_matrix
+        else:
+            self.cat_matrix = np.array([
+                [cat_dissimilarity_matrix(cat1, cat2) for cat1 in categories] for cat2 in categories
+            ])
 
         # sanity checks on the categorical_dissimilarity_matrix
         assert isinstance(self.cat_matrix, np.ndarray)
@@ -140,7 +150,7 @@ class CategoricalDissimilarity(AbstractDissimilarity):
                         f"In segment {unit.segment} for annotator {annotator}: "
                         f"annotation of category {unit.category} is not in "
                         f"set {set(self.categories_dict.keys())} of allowed categories")
-            cat_array[-1] = -1 # We add an empty unit at the end of each line so that the carthesian product uses it
+            cat_array[-1] = -1  # We add an empty unit at the end of each line so that the carthesian product uses it
             categories_arrays.append(cat_array)
         return categories_arrays
 
@@ -195,25 +205,31 @@ class CategoricalDissimilarity(AbstractDissimilarity):
         plt.show()
 
     @staticmethod
-    @nb.njit(nb.float32(nb.int32[:, :],
-                        nb.types.ListType(nb.float32[:, ::1]),
-                        nb.types.ListType(nb.int16[::1]),
-                        nb.float32,
-                        nb.float32[:, :]))
+    @nb.njit(nb.float32[:](nb.int32[:, :],
+                           nb.types.ListType(nb.int16[::1]),
+                           nb.float32,
+                           nb.float32[:, :]))
     def alignments_disorder(units_tuples_ids: np.ndarray,
-                            units_positions: List[np.ndarray],
                             units_categories: List[np.ndarray],
                             delta_empty: np.float32,
[…]
