### 2021-05-28 5033352
new cli argument for different categorical dissimilarities & bug fixes
