### 2021-05-28 73419b6
fix of wrong normalization of categorical dissimilarity matrix
