### 2021-05-31 7af0557
resolved dependencies, reformatted some code, added saving an alignement png to a file
