+++ pygamma_agreement/gen_continuum.py
@@ -1,0 +1,2 @@
+from .continuum import Continuum
+
