+++ requirements.txt
@@ -7,5 +7,10 @@ numba>= 0.48.0
 typing_extensions>= 3.7.4.3
 TextGrid>=1.5
 pympi-ling>=1.69
-pyannote.database >= 4.0.1
+pyannote.database >= 4.1
 dataclasses >= 0.7; python_version <'3.7'
+python-Levenshtein >= 0.12.2
+pytest~=6.2.4
+matplotlib~=3.4.2
+setuptools~=57.0.0
+ipython~=7.24.0
