### 2021-05-31 fa3f28f
class for random continuum generation, some imports
