### 2021-05-31 30cf3f5
functions to export plots in files, added options to display categories in plots
