### 2021-06-01 2a08647
small tweak of annotation texts
