### 2021-06-01 33d0475
sets the best alignment calculted by compute_gamma's attribute to self
