### 2021-06-01 4044ad0
using average duration of reference units for creation of false-positives
