### 2021-06-02 8fbf51b
function for plotting a continuum
