### 2021-06-02 701873c
some properties for statistical values used for the CST
