+++ pygamma_agreement/cst.py
@@ -1,3 +1,4 @@
+import math
 import random
 
 import numpy.random
@@ -6,6 +7,7 @@ import logging
 from .continuum import Continuum, Unit
 from typing import Union, Iterable, List, Callable, Set, Tuple
 from sortedcontainers import SortedDict, SortedSet
+from multiprocessing import pool
 import numpy as np
 from pyannote.core import Segment
 
@@ -16,9 +18,9 @@ class CorpusShufflingTool:
     (https://www.aclweb.org/anthology/J15-3003.pdf#page=30).
     Beware that the reference continuum is a copy of the given continuum.
     """
-    SHIFT_FACTOR = 0.5  # must be < 1 or there might be problems of empty ranges
-    SPLIT_FACTOR = 5
-    FALSE_POS_FACTOR = 5
+    SHIFT_FACTOR = 2
+    SPLIT_FACTOR = 1
+    FALSE_POS_FACTOR = 1
 
     def __init__(self,
                  magnitude: float,
@@ -38,83 +40,144 @@ class CorpusShufflingTool:
         """
         self._seed: int = seed
         self.magnitude: float = magnitude
+        assert len(reference_continuum.annotators) == 1
         self._reference_continuum: Continuum = reference_continuum
         self._categories = self._reference_continuum.categories.union(categories)
 
-    def corpus_shuffle(self,
-                       new_annotators: Union[int, Iterable[str]],
-                       reference_annotator: str = None,
-                       include_reference=False):
-        """
-        Generates a shuffled corpus with the provided (or generated) reference annotation set,
-        using the method described in 6.3 of @gamma-paper, https://www.aclweb.org/anthology/J15-3003.pdf#page=30
-        """
+    def corpus_from_reference(self, new_annotators: Union[int, Iterable[str]]):
         continuum = Continuum()
-        reference_annotators = self._reference_continuum.annotators
-
-        if reference_annotator is None:
-            reference_annotator = next(iter(reference_annotators))
-            if len(reference_annotators) > 1:
-                logging.warning("CST was given a multi-annotator reference, but reference annotator was not\n"
-                                "specified. A default annotator (1st in alphabetical) was used.")
-        else:
-            assert (reference_annotator in reference_annotators)
-
-        units = self._reference_continuum[reference_annotator]
-        if include_reference:
-            continuum[reference_annotator] = units
-
+        annotator_set = iter(self._reference_continuum.annotators)
         if isinstance(new_annotators, int):
-            new_annotators = (f"annotator_cst_{i}" for i in range(new_annotators))
-
-        shift_max = self.magnitude * self.SHIFT_FACTOR
-        bounds_inf, bounds_sup = (next(iter(units)).segment.start, next(reversed(units)).segment.end)
-        avg_dur_false_pos = np.average([unit.segment.end - unit.segment.start for unit in units])
-        for new_annotator in new_annotators:
-            assert new_annotator not in reference_annotators
-            continuum.add_annotator(new_annotator)
-            for unit in units:
-                # false negatives
-                if np.random.uniform() < self.magnitude:
-                    continue
-                # category TODO
-                category = unit.annotation
-                # positions
-                len_unit = unit.segment.end - unit.segment.start
+            new_annotators = SortedSet(f"annotator_{i}" for i in range(new_annotators))
+        else:
+            new_annotators = SortedSet(new_annotators)
+        for unit in self._reference_continuum[next(annotator_set)]:
+            for new_annotator in new_annotators:
                 continuum.add(new_annotator,
-                              Segment(unit.segment.start + np.random.uniform(-shift_max*len_unit/2,
-                                                                             shift_max*len_unit/2),
-                                      unit.segment.end + np.random.uniform(-shift_max*len_unit/2,
[…]
