### 2021-06-03 66496b0
converted lists to sets to ensure equality if order of elements is not the same
