### 2021-06-03 2143e77
minor correction for computation of weights of each annotation
