### 2021-06-03 0744cb8
minor modifications & comments
