### 2021-06-03 b00f503
plotting tool for benchmarking gamma with the cst
