### 2021-06-04 a50f1aa
additionnal tests for the corpus shuffling tool
