### 2021-06-04 651f7b1
documentation, code factorisation, typing, small fixes
