+++ pygamma_agreement/cst.py
@@ -30,9 +30,9 @@ import math
 import numpy.random
 import logging
 
-from .continuum import Continuum, Unit
+from .continuum import *
 from .dissimilarity import AbstractDissimilarity
-from typing import Union, Iterable, Callable, List, Tuple
+from typing import Union, Iterable, Callable, List, Tuple, Optional
 from sortedcontainers import SortedSet
 import numpy as np
 from pyannote.core import Segment
@@ -42,7 +42,6 @@ class CorpusShufflingTool:
     """
     Corpus shuffling tool as detailed in section 6.3 of @gamma-paper
     (https://www.aclweb.org/anthology/J15-3003.pdf#page=30).
-    Beware that the reference continuum is a copy of the given continuum.
     """
     SHIFT_FACTOR = 2
     SPLIT_FACTOR = 1
@@ -51,7 +50,7 @@ class CorpusShufflingTool:
     def __init__(self,
                  magnitude: float,
                  reference_continuum: Continuum,
-                 seed: int = 4772,
+                 seed: Optional[float] = 4772,
                  categories: Iterable[str] = None):
         """
         Parameters
@@ -64,7 +63,7 @@ class CorpusShufflingTool:
             this is used to consider additionnal categories when shuffling the corpus, in the eventuality that the
             reference continuum does not contain any unit of a possible category.
         """
-        self._seed: int = seed
+        self._seed: Optional[int] = seed
         self.magnitude: float = magnitude
         reference_annotators = reference_continuum.annotators
         if len(reference_annotators) > 1:
@@ -76,6 +75,10 @@ class CorpusShufflingTool:
         if categories is not None:
             for category in categories:
                 self._categories.add(category)
+
+    def __setseed(self):
+        if self._seed is not None:
+            np.random.seed(self._seed)
 
     def corpus_from_reference(self, new_annotators: Union[int, Iterable[str]]):
         continuum = Continuum()
@@ -95,8 +98,9 @@ class CorpusShufflingTool:
         Tweaks the continuum by shifting the ends of each segment, with uniformly distributed values
         of bounds proportionnal to the magnitude of the CST and the length of the segment.
         """
+        self.__setseed()
         shift_max = self.magnitude * self.SHIFT_FACTOR * \
-            self._reference_continuum.avg_length_unit
+                    self._reference_continuum.avg_length_unit
         for annotator in continuum.annotators:
             for unit in continuum[annotator]:
                 continuum[annotator].remove(unit)
@@ -112,6 +116,7 @@ class CorpusShufflingTool:
         Every unit (for each annotator) have a probability equal to the magnitude of being removed.
         If this probability is one, a single random unit (for each annotator) will be left alone.
         """
+        self.__setseed()
         for annotator in continuum.annotators:
             security = np.random.choice(continuum[annotator])
             # security : if an annotator doesnt have any annotations gamma cant be computed.
@@ -129,6 +134,7 @@ class CorpusShufflingTool:
         The length of the segment is random (normal distribution) based on the average and standard deviation
         of those of the reference.
         """
+        self.__setseed()
         ref_units = self._reference_continuum[self._reference_annotator]
         avg_dur = np.average([unit.segment.end - unit.segment.start for unit in ref_units])
         var_dur = np.std([unit.segment.end - unit.segment.start for unit in ref_units])
@@ -148,9 +154,8 @@ class CorpusShufflingTool:
                          overlapping_fun: Callable[[str, str], float] = None,
                          prevalence: bool = False):
         """
-        Shuffles the category using the process described in section 3.3.5 of
-        https://hal.archives-ouvertes.fr/hal-00769639/.
-
+        Shuffles the categories in the annotations using the process described in
[…]
