### 2021-06-06 bootphon/pygamma-agreement#24 @faroit
[JOSS] Improve clarity of statement of need section
Hi authors, I went once more through the paper and I have three more comments that would be great if those can be addressed in another iteration:

* [x] the introduction reads very well but misses out the technical details of how the gamma-measure actually works. I don't think there would be enough space to include more details on the algorithm but if possible, there could be 1 or 2 sentences added in shortly (and intuitively) to describe the basic idea of the gamma agreement algorithm.
* [x] there are already a good amount of references when it comes to the comparison of the measures. I would however welcome one more reference of a well-cited study that is using the gamma agreement measure for their analysis of results.
* [x] related to #23, the plots are now fixed but I would welcome a more descriptive figure caption that describes what the continuum and the alignment is showing in plain words.

This issue is part of my JOSS review https://github.com/openjournals/joss-reviews/issues/2989
