### 2021-06-06 #2989 @faroit
@hadware I checked the last remaining issues for this review. I think this is looking good. Can you please address the minor issues I raised in https://github.com/bootphon/pygamma-agreement/issues/24 before we can proceed?
