### 2021-06-07 bootphon/pygamma-agreement#24 @hadware
Thanks for the feedback, i'll update the paper to include your remarks.
