### 2021-06-07 a879a6f
test with use of sets of categories instead of lists
