### 2021-06-07 5c31ef8
1st version of gamma-cat/k-disorder
