### 2021-06-07 07ce14d
1st version of gamma-cat/k-disorder and prep work
