+++ pygamma_agreement/continuum.py
@@ -33,6 +33,7 @@ Continuum and corpus
 """
 import csv
 import logging
+import os
 import random
 from copy import deepcopy
 from functools import total_ordering
@@ -54,7 +55,7 @@ from .numba_utils import chunked_cartesian_product
 if TYPE_CHECKING:
     from .alignment import UnitaryAlignment, Alignment
 
-CHUNK_SIZE = 2 ** 23
+CHUNK_SIZE = 2 ** 25 // os.cpu_count()
 
 # defining Annotator type
 Annotator = str
@@ -514,10 +515,8 @@ class Continuum:
         """
         return iter(self._annotations)
 
-
     def compute_disorders(self, dissimilarity: AbstractDissimilarity):
         assert len(self.annotators) >= 2, "Disorder cannot be computed with less than two annotators."
-
 
         disorder_args = dissimilarity.build_args(self)
 
@@ -525,7 +524,6 @@ class Continuum:
         for annotator, arr in self._annotations.items():
             assert len(arr) > 0, f"Disorder cannot be computed because annotator {annotator} has no annotations."
             nb_unit_per_annot.append(len(arr) + 1)
-
 
         all_disorders = []
         all_valid_tuples = []
@@ -645,15 +643,17 @@ class Continuum:
         p = Pool()
         result_pool = [
             # Step one : computing the disorders of a batch of random samples from the continuum (done in parallel)
-            p.apply_async(_compute_disorder_job,
+            p.apply_async(_compute_best_alignment_job,
                           (Continuum.sample_from_continuum(self, pivot_type, ground_truth_annotators),
                            dissimilarity,)
                           )
             for _ in range(n_samples)]
-        chance_disorders = []
+        chance_best_alignments: List[Alignment] = []
+        chance_disorders: List[float] = []
         logging.info(f"Starting computation for a batch of {n_samples} random samples...")
         for result in result_pool:
-            chance_disorders.append(result.get())
+            chance_best_alignments.append(result.get())
+            chance_disorders.append(chance_best_alignments[-1].disorder)
         logging.info("done.")
 
         if precision_level is not None:
@@ -669,7 +669,7 @@ class Continuum:
             logging.debug(f"Number of required samples for confidence {precision_level}: {required_samples}")
             if required_samples > n_samples:
                 result_pool = [
-                    p.apply_async(_compute_disorder_job,
+                    p.apply_async(_compute_best_alignment_job,
                                   (Continuum.sample_from_continuum(self, pivot_type, ground_truth_annotators),
                                    dissimilarity,)
                                   )
@@ -678,7 +678,7 @@ class Continuum:
                 logging.info(f"Computing second batch of {required_samples - n_samples} "
                              f"because variation was too high.")
                 for i, result in enumerate(result_pool):
-                    chance_disorders.append(result.get())
+                    chance_best_alignments.append(result.get())
                 logging.info("done.")
 
         p.close()
@@ -690,8 +690,9 @@ class Continuum:
             best_alignment=best_alignment,
             pivot_type=pivot_type,
             n_samples=n_samples,
-            chance_disorders=np.array(chance_disorders),
-            precision_level=precision_level
+            chance_alignments=chance_best_alignments,
+            precision_level=precision_level,
+            dissimilarity=dissimilarity
         )
 
     def compute_gamma_cat(self):
@@ -721,12 +722,14 @@ class Continuum:
 @dataclass
 class GammaResults:
     """
-    Gamma results object. Stores information about a gamma measure computation.
+    Gamma results object. Stores the informations about a gamma measure computation,
+    used for getting the values of measures from the gamma family (gamma, gamma-cat and gamma-k).
     """
[…]
