### 2021-06-07 3304701
d_pos and d_cat functions for use in gamma-cat
