### 2021-06-08 20db104
function for plotting an alignment
