### 2021-06-08 a0ecb87
corrected a small difference in the disorder of an alignment (normalized by number of unitary alignments instead of mean of annotations per annotator)
