### 2021-06-08 daaaf32
Saving an additionnal gamma computation per test
