### 2021-06-08 4ba19a6
Modified weighting of a couple of units to match the gamma software
