### 2021-06-09 2f5b239
Normalization of disorder of an alignment by mean number of annotation per annotator for right value of gamma
