### 2021-06-09 b297316
Update of tests with right values of disorder/gamma after a normalization error correction
