+++ tests/test_alignement.py
@@ -8,7 +8,7 @@ from pygamma_agreement.alignment import SetPartitionError
 from pygamma_agreement.alignment import UnitaryAlignment, Alignment
 from pygamma_agreement.continuum import Continuum, Unit
 from pygamma_agreement.dissimilarity import CombinedCategoricalDissimilarity
-
+from sortedcontainers import SortedSet
 
 def test_alignment_checking():
     continuum = Continuum()
@@ -24,9 +24,8 @@ def test_alignment_checking():
     # checking valid alignment
     alignment = Alignment([], continuum=continuum)
 
-    n_tuple = (('liza', Unit(Segment(1, 5), 'Carol')),
-               ('pierrot', Unit(Segment(2, 6), 'Carol'))
-               )
+    n_tuple = [('liza', Unit(Segment(1, 5), 'Carol')),
+               ('pierrot', Unit(Segment(2, 6), 'Carol'))]
     unitary_alignment = UnitaryAlignment(n_tuple)
     alignment.unitary_alignments.append(unitary_alignment)
 
@@ -34,9 +33,8 @@ def test_alignment_checking():
     with pytest.raises(SetPartitionError):
         alignment.check(continuum)
 
-    n_tuple = (('liza', Unit(Segment(6, 8), 'Bob')),
-               ('pierrot', None)
-               )
+    n_tuple = [('liza', Unit(Segment(6, 8), 'Bob')),
+               ('pierrot', None)]
     unitary_alignment = UnitaryAlignment(n_tuple)
     alignment.unitary_alignments.append(unitary_alignment)
 
@@ -44,9 +42,8 @@ def test_alignment_checking():
     alignment.check(continuum)
 
     # checking with extra tuple
-    n_tuple = (('liza', Unit(Segment(6, 8), 'Bob')),
-               ('pierrot', None)
-               )
+    n_tuple = [('liza', Unit(Segment(6, 8), 'Bob')),
+               ('pierrot', None)]
     unitary_alignment = UnitaryAlignment(n_tuple)
     alignment.unitary_alignments.append(unitary_alignment)
 
@@ -54,8 +51,9 @@ def test_alignment_checking():
     with pytest.raises(SetPartitionError):
         alignment.check(continuum)
 
+
 def test_unitary_alignment():
-    categories = ['Carol', 'Bob', 'Alice', 'Jeremy']
+    categories = SortedSet(['Carol', 'Bob', 'Alice', 'Jeremy'])
     cat = np.array([[0, 0.5, 0.3, 0.7],
                     [0.5, 0., 0.6, 0.4],
                     [0.3, 0.6, 0., 0.7],
@@ -65,14 +63,15 @@ def test_unitary_alignment():
         delta_empty=0.5,
         cat_dissimilarity_matrix=cat,
         alpha=1)
-    n_tuple = (('liza', Unit(Segment(12, 18), "Carol")),
+    n_tuple = [('liza', Unit(Segment(12, 18), "Carol")),
                ('pierrot', Unit(Segment(12, 18), "Alice")),
-               ('hadrien', None))
+               ('hadrien', None)]
     unitary_alignment = UnitaryAlignment(n_tuple)
 
     assert (unitary_alignment.compute_disorder(combi_dis)
            ==
-           pytest.approx(0.383, 0.001))
+           pytest.approx(0.574999988079071, 0.001))
+
 
 def test_alignment():
     continuum = Continuum()
@@ -99,7 +98,7 @@ def test_alignment():
 
     continuum.add_annotation('hadrien', annotation)
 
[…]
