### 2021-06-09 66b6199
Added test for mean_nb_units_per_annotators
