### 2021-06-09 a1ccca1
Added a small summary of how the gamma works. Added a study making use of the gamma. Added clarifications on the plots. #24
