### 2021-06-09 bootphon/pygamma-agreement#24 @hadware
I updated the paper to include your suggestions. (point 1 and 2 are in the Statements of needs section, point 3 was added as additional explanations under the respective figures). I hope this made the paper clearer.
