### 2021-06-09 bootphon/pygamma-agreement#24 @faroit
@hadware thanks a lot. This greatly improves the readability of the paper. I just made a minor comment on a1ccca1 feel free to close this issue (with our without incorporating my proposal)
