### 2021-06-09 22e0a82
Small corrections regarding comments from @faroit
