+++ paper/paper.md
@@ -39,7 +39,7 @@ The Gamma ($\gamma$) Inter-Annotator Agreement Measure was proposed by [@gamma-p
 This quantification will have to satisfy some constraints: segmentation, unitizing, categorization, weighted categorization and the support for any number of annotators. They should also provide a chance-corrected value.
 Other measures, such as the $\kappa$ [@kappa-paper] or Krippendorff's $\alpha$'s [@alpha-paper], have existed for some time to deal with these constraints, but never could address all of them at once. A detailed comparison between metrics is available in [@gamma-paper].
 
-To solve all of these constraints at once, the $\gamma$-agreement works in 3 steps:  1) a _disorder_ (i.e. a cost) is computed for each potential alignments between the different annotators's units, using an annotation-dependent _dissimilarity_ (akin to a distance between units). This _disorder_ models the disagreement between annotators. 2) Using a convex optimization algorithm, a global alignment with the lowest total disorder is found. 3) By sampling random and synthetic annotations from the original annotation and computing their own disorders, the original annotation's disorder value is chance-corrected to provide the final $\gamma$-agreement measure.
+To solve all of these constraints at once, the $\gamma$-agreement works in three steps:  1) a _disorder_ (i.e. a cost) is computed for each potential alignments between the different annotators' units, using an annotation-dependent _dissimilarity_ (akin to a distance between units). This _disorder_ models the disagreement between annotators. 2) Using a convex optimization algorithm, a global alignment with the lowest total disorder is found. 3) By sampling random and synthetic annotations from the original annotation and computing their own disorders, the original annotation's disorder value is chance-corrected to provide the final $\gamma$-agreement measure.
 The authors of [@gamma-paper] [provided a Java freeware](https://gamma.greyc.fr/) GUI implementation along with their paper. This implementation has already been used by some researchers  [@gamma-usage] to compute an inter-rater agreement measure on their annotations.
 
 However, linguist and automated speech researchers today use analysis pipelines that are either Python or shell scripts. 
