### 2021-06-09 bootphon/pygamma-agreement#24 @hadware
Thanks for pickup on those again! :) I corrected these small issues in 22e0a82761404c06b7323bc33dabf45fc2256c40 .
