### 2021-06-09 #2989 @hadware
All right:

10.5281/zenodo.4917244

Here's the ZENODO DOI, and here's the package's page : https://zenodo.org/record/4917244 .
I hope this fits the bill.
