### 2021-06-09 7f90f59
skipping null units when computing gammacat to match the GammaSoftware implementation
