### 2021-06-09 f965d64
removed random seed in test
