### 2021-06-10 83dc0b1
Added a way pretty plot of an alignment of annotations
