### 2021-06-10 #2989 @kthyng
@hadware a few comments on your paper:

* is this how you usually display your affiliations?
* first sentence under Statement of Need: the reference shouldn't be in parentheses — remove the [] around the reference.
* Same with the final sentence in that paragraph and several more times in the paper. Please update all of these.
