### 2021-06-10 #2989 @hadware
Hey @kthyng !

The current version (0.1.6) is good :)

Sorry for the refs, i thought the brackets were mandatory to render a citation. I'll update those and tell you when it's done.
