### 2021-06-11 21ea375
changes in continuum sampler with reproducing the GammaSoftware results in mind
