+++ pygamma_agreement/continuum.py
@@ -60,6 +60,7 @@ CHUNK_SIZE = 2 ** 25 // os.cpu_count()
 # defining Annotator type
 Annotator = str
 PivotType = Literal["float_pivot", "int_pivot"]
+BoundsType = Literal["wide", "narrow"]
 PrecisionLevel = Literal["high", "medium", "low"]
 
 # percentages for the precision
@@ -172,7 +173,9 @@ class Continuum:
 
     @classmethod
     def sample_from_continuum(cls, continuum: 'Continuum',
-                              pivot_type: PivotType = "float_pivot",
+                              pivot_type: PivotType = "int_pivot",
+                              min_dist_between_pivots: Optional[float] = None,
+                              bounds_type: BoundsType = "wide",
                               ground_truth_annotators: Optional[List[Annotator]] = None) -> 'Continuum':
         """Generate a new random annotation from a single continuum
                 Strategy from figure 12
@@ -181,25 +184,44 @@ class Continuum:
                 ... <pygamma_agreement.continuum.Continuum at 0x7f5527a19588>
                 """
         assert pivot_type in ('float_pivot', 'int_pivot')
+        assert bounds_type in ('wide', 'narrow')
+        min_dist_between_pivots = continuum.avg_length_unit / 2
 
         last_start_time = max(unit.segment.start for _, unit in continuum)
+        if bounds_type == 'wide':
+            bound_inf, bound_sup = continuum.bounds
+        else:
+            bound_inf, bound_sup = continuum.bounds_narrow
         new_continuum = Continuum()
         if ground_truth_annotators is not None:
             assert set(continuum.annotators).issuperset(set(ground_truth_annotators))
             annotators = ground_truth_annotators
         else:
             annotators = continuum.annotators
-
         # TODO: why not sample from the whole continuum?
         # TODO : shouldn't the sampled annotators nb be equal to the annotators amount?
+        pivots = []
         for idx in range(continuum.num_annotators):
             if pivot_type == 'float_pivot':
-                pivot = random.uniform(continuum.avg_length_unit, last_start_time)
-            else:
-                pivot = random.randint(np.floor(continuum.avg_length_unit),
-                                       np.ceil(last_start_time))
+                pivot: float = random.uniform(bound_inf, bound_sup)
+                if min_dist_between_pivots is not None:
+                    # While the pivot is closer than min_dist to a precedent pivot, pick another one
+                    # (takes wrapping of continuum into consideration).
+                    while any(map((lambda x: abs(x - pivot) <= min_dist_between_pivots or
+                                   abs(x - (pivot - last_start_time)) <= min_dist_between_pivots),
+                                  pivots)):
+                        pivot = random.uniform(bound_inf, bound_sup)
 
-            rnd_annotator = random.choice(annotators)
+            else:
+                pivot: int = random.randint(np.floor(bound_inf), np.ceil(bound_sup))
+                if min_dist_between_pivots is not None:
+                    while any(map((lambda x: abs(x - pivot) <= min_dist_between_pivots or
+                                   abs(x - (pivot - last_start_time)) <= min_dist_between_pivots),
+                                  pivots)):
+                        pivot = random.randint(np.floor(bound_inf), np.ceil(bound_sup))
+            pivots.append(pivot)
+
+            rnd_annotator = random.choice(list(annotators))
             units = continuum._annotations[rnd_annotator]
             sampled_annotation = SortedSet()
             for unit in units:
@@ -216,7 +238,7 @@ class Continuum:
     def __init__(self, uri: Optional[str] = None):
         self.uri = uri
         # Structure {annotator -> SortedSet[Unit]}
[…]

[…]
