### 2021-06-11 218571c
paper: slight fix to the bullet points
