### 2021-06-12 #2989 @hadware
Thanks to all for your help and feedback with this paper :) I've already volunteered to be a reviewer to JOSS, i'll be waiting for new packages to review :)
