### 2021-06-13 #2989 @apiad
> @faroit, @apiad – many thanks for your reviews here! JOSS relies upon the volunteer efforts of folks like you and we simply wouldn't be able to do this without you
>
> @hadware – your paper is now accepted and published in JOSS

It has been my pleasure! Count me on whenever you need, I'll do my best to be available :smiley:
